change json schema and language bindings for opt progress
